New test case and test re-org
