+++ packages/firestore/test/unit/specs/collection_spec.test.ts
@@ -15,182 +15,12 @@
  * limitations under the License.
  */
 
-import { newQueryForPath } from '../../../src/core/query';
-import { deletedDoc, doc, filter, query } from '../../util/helpers';
+import { doc, query } from '../../util/helpers';
 
 import { describeSpec, specTest } from './describe_spec';
 import { spec } from './spec_builder';
 
 describeSpec('Collections:', [], () => {
-  specTest('8474', [], () => {
-    // onSnapshot(fullQuery)
-    const fullQuery = query('collection');
-
-    // getDocs(filterQuery)
-    const filterQuery = query('collection', filter('included', '==', true));
-
-    const docA = doc('collection/a', 1000, { key: 'a', included: false });
-    const docA2 = doc('collection/a', 1007, { key: 'a', included: true });
-    const docC = doc('collection/c', 1002, { key: 'c', included: true });
-
-    const limboQueryC = newQueryForPath(docC.key.path);
-
-    return (
-      spec()
-        // onSnapshot(fullQuery) - fullQuery is listening to documents in the collection for the full test
-        .userListens(fullQuery)
-        .watchAcksFull(fullQuery, 1001, docA, docC)
-        .expectEvents(fullQuery, {
-          fromCache: false,
-          added: [docA, docC]
-        })
-
-        // docC was deleted, this puts docC in limbo and causes a snapshot from cache (metadata-only change)
-        .watchSends({ removed: [fullQuery] }, docC)
-        .watchCurrents(fullQuery, 'resume-token-1002')
-        .watchSnapshots(1002)
-        .expectLimboDocs(docC.key)
-        .expectEvents(fullQuery, {
-          fromCache: true
-        })
-
-        // User begins getDocs(filterQuery)
-        .userListensForGet(filterQuery)
-
-        // getDocs(filterQuery) will not resolve on the snapshot from cache
-        .expectEvents(filterQuery, {
-          fromCache: true,
-          added: [docC]
-        })
-
-        // Watch acks limbo and filter queries
-        .watchAcks(limboQueryC)
-        .watchAcks(filterQuery)
-
-        // Watch responds to limboQueryC - docC was deleted
-        .watchSends({ affects: [limboQueryC] })
-        .watchCurrents(limboQueryC, 'resume-token-1009')
-        .watchSnapshots(1009, [limboQueryC, fullQuery])
-
-        // However, docC is still in limbo because there has not been a global snapshot
-        .expectLimboDocs(docC.key)
-
-        // Rapid events of document update and delete caused by application
-        .watchSends({ removed: [filterQuery] }, docA)
-        .watchCurrents(filterQuery, 'resume-token-1004')
-        .watchSends({ affects: [filterQuery] }, docC)
-        .watchCurrents(filterQuery, 'resume-token-1005')
-        .watchSends({ removed: [filterQuery] }, docC)
-        .watchSends({ affects: [filterQuery] }, docA2)
-        .watchCurrents(filterQuery, 'resume-token-1007')
-
-        .watchSnapshots(1010, [fullQuery, limboQueryC])
-
-        // All changes are current and we get a global snapshot
-        .watchSnapshots(1010, [])
-
-        // Now docC is out of limbo
-        .expectLimboDocs()
-        .expectEvents(fullQuery, {
-          fromCache: false,
-          modified: [docA2],
-          removed: [docC]
-        })
-        // Now getDocs(filterQuery) can be resolved
-        .expectEvents(filterQuery, {
-          fromCache: false,
-          removed: [docC],
-          added: [docA2]
-        })
-    );
-  });
-  specTest('8474-deleted', [], () => {
-    // onSnapshot(fullQuery)
-    const fullQuery = query('collection');
-
-    // getDocs(filterQuery)
-    const filterQuery = query('collection', filter('included', '==', true));
-
-    const docA = doc('collection/a', 1000, { key: 'a', included: false });
-    const docA2 = doc('collection/a', 1007, { key: 'a', included: true });
-    const docC = doc('collection/c', 1002, { key: 'c', included: true });
-    const docCDeleted = deletedDoc('collection/c', 1009);
-
-    const limboQueryC = newQueryForPath(docC.key.path);
-
-    return (
-      spec()
-        // onSnapshot(fullQuery) - fullQuery is listening to documents in the collection for the full test
-        .userListens(fullQuery)
-        .watchAcksFull(fullQuery, 1001, docA, docC)
-        .expectEvents(fullQuery, {
-          fromCache: false,
-          added: [docA, docC]
-        })
-
-        // docC was deleted, this puts docC in limbo and causes a snapshot from cache (metadata-only change)
-        .watchSends({ removed: [fullQuery] }, docC)
-        .watchCurrents(fullQuery, 'resume-token-1002')
-        .watchSnapshots(1002)
-        .expectLimboDocs(docC.key)
-        .expectEvents(fullQuery, {
-          fromCache: true
-        })
-
-        // User begins getDocs(filterQuery)
-        .userListensForGet(filterQuery)
-
-        // getDocs(filterQuery) will not resolve on the snapshot from cache
-        .expectEvents(filterQuery, {
-          fromCache: true,
-          added: [docC]
-        })
-
-        // Watch acks limbo and filter queries
-        .watchAcks(limboQueryC)
-        .watchAcks(filterQuery)
-
-        // Watch responds to limboQueryC - docC was deleted
-        .watchSends({ affects: [limboQueryC] }, docCDeleted)
-        .watchCurrents(limboQueryC, 'resume-token-1009')
-        .watchSnapshots(1009, [limboQueryC, fullQuery])
-
-        // However, docC is still in limbo because there has not been a global snapshot
-        .expectLimboDocs(docC.key)
-
-        // Rapid events of document update and delete caused by application
-        .watchSends({ removed: [filterQuery] }, docA)
-        .watchCurrents(filterQuery, 'resume-token-1004')
-        .watchSends({ affects: [filterQuery] }, docC)
-        .watchCurrents(filterQuery, 'resume-token-1005')
-        .watchSends({ removed: [filterQuery] }, docC)
-        .watchSends({ affects: [filterQuery] }, docA2)
-        .watchCurrents(filterQuery, 'resume-token-1007')
-
-        .watchSnapshots(1010, [fullQuery, limboQueryC])
-
-        // All changes are current and we get a global snapshot
-        .watchSnapshots(1010, [])
-
-        // Now docC is out of limbo
-        .expectLimboDocs()
-        .expectEvents(fullQuery, {
-          fromCache: false,
-          modified: [docA2],
-          removed: [docC]
-        })
-        // Now getDocs(filterQuery) can be resolved
-        .expectEvents(filterQuery, {
-          fromCache: false,
-          removed: [docC],
-          added: [docA2]
-        })
-
-        // No more expected events
-        .watchSnapshots(1100, [])
-    );
-  });
-
   specTest('Events are raised after watch ack', [], () => {
     const query1 = query('collection');
     const doc1 = doc('collection/key', 1000, { foo: 'bar' });
